% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.14_function_comparative_habitat_selection.R
\name{analyze_single_dataset}
\alias{analyze_single_dataset}
\title{Analyze a single presence-absence dataset}
\usage{
analyze_single_dataset(
  data,
  dataset_name,
  formula,
  sample_size,
  ntree,
  depth_variable,
  fish_id_variable,
  time_variable,
  type_variable,
  n_depth_points,
  create_plots
)
}
\description{
Analyze a single presence-absence dataset
}
