% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{calculate_95_ellipse_area}
\alias{calculate_95_ellipse_area}
\title{Calculate 95\% confidence ellipse area}
\usage{
calculate_95_ellipse_area(x_coords, y_coords)
}
\description{
Calculate 95\% confidence ellipse area
}
