% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{calculate_constrained_convex_hull_area}
\alias{calculate_constrained_convex_hull_area}
\title{Updated constrained convex hull calculation to work with reference raster}
\usage{
calculate_constrained_convex_hull_area(
  x_coords,
  y_coords,
  grid_resolution,
  reference_raster = NULL
)
}
\description{
Updated constrained convex hull calculation to work with reference raster
}
