% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{calculate_constrained_convex_hull_area_with_data}
\alias{calculate_constrained_convex_hull_area_with_data}
\title{Calculate constrained convex hull with spatial data}
\usage{
calculate_constrained_convex_hull_area_with_data(
  x_coords,
  y_coords,
  grid_resolution,
  reference_raster = NULL
)
}
\description{
Calculate constrained convex hull with spatial data
}
