% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{calculate_convex_hull_area}
\alias{calculate_convex_hull_area}
\title{Calculate convex hull area}
\usage{
calculate_convex_hull_area(x_coords, y_coords)
}
\description{
Calculate convex hull area
}
