% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{calculate_grid_cell_usage}
\alias{calculate_grid_cell_usage}
\title{Updated grid cell usage calculation to work with reference raster}
\usage{
calculate_grid_cell_usage(
  x_coords,
  y_coords,
  grid_resolution,
  reference_raster = NULL
)
}
\description{
Updated grid cell usage calculation to work with reference raster
}
