% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{calculate_grid_cell_usage_with_data}
\alias{calculate_grid_cell_usage_with_data}
\title{Calculate grid cell usage with spatial data}
\usage{
calculate_grid_cell_usage_with_data(
  x_coords,
  y_coords,
  grid_resolution,
  reference_raster = NULL
)
}
\description{
Calculate grid cell usage with spatial data
}
