% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{calculate_group_space_use}
\alias{calculate_group_space_use}
\title{Calculate space use metrics for a single group}
\usage{
calculate_group_space_use(
  group_data,
  methods,
  grid_resolution,
  reference_raster = NULL
)
}
\description{
Calculate space use metrics for a single group
}
