% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{calculate_space_use_zoom}
\alias{calculate_space_use_zoom}
\title{Calculate zoom extent for map considering both track data and space use area}
\usage{
calculate_space_use_zoom(
  track_points = NULL,
  space_use_spatial_data = NULL,
  method = NULL,
  buffer_factor = 0.15
)
}
\description{
Calculate zoom extent for map considering both track data and space use area
}
