% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/1.7.11_function_space_use_point_generation_part2.R
\name{compare_space_use_methods}
\alias{compare_space_use_methods}
\title{Compare multiple space use methods by generating points for each}
\usage{
compare_space_use_methods(
  space_use_results,
  track_data,
  methods = c("convex_hull", "constrained_convex_hull", "bounding_box"),
  n_points = 100,
  fish_ids = NULL,
  time_periods = NULL,
  reference_raster = NULL
)
}
\arguments{
\item{space_use_results}{Output from calculate_space_use function}

\item{track_data}{Original track data}

\item{methods}{Vector of methods to compare}

\item{n_points}{Number of points per method}

\item{fish_ids}{Which fish to include}

\item{time_periods}{Which time periods to include}

\item{reference_raster}{Optional reference raster}
}
\value{
A data frame containing generated points for each method with comparison metadata
}
\description{
Compare multiple space use methods by generating points for each
}
