% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.14_function_comparative_habitat_selection.R
\name{create_comparison_analysis}
\alias{create_comparison_analysis}
\title{Create comparison analysis between positioning and space use results}
\usage{
create_comparison_analysis(
  positioning_results,
  space_use_results,
  depth_variable
)
}
\description{
Create comparison analysis between positioning and space use results
}
