% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{create_constrained_grid_raster_data}
\alias{create_constrained_grid_raster_data}
\title{Create constrained hull by finding which depth_raster cells are within convex hull polygon}
\usage{
create_constrained_grid_raster_data(track_points, reference_raster)
}
\description{
Create constrained hull by finding which depth_raster cells are within convex hull polygon
}
