% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{create_ellipse_polygon}
\alias{create_ellipse_polygon}
\title{Create ellipse polygon for 95\% confidence ellipse}
\usage{
create_ellipse_polygon(x_coords, y_coords, n_points = 100)
}
\description{
Create ellipse polygon for 95\% confidence ellipse
}
