% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{create_ellipse_polygon_data}
\alias{create_ellipse_polygon_data}
\title{Create ellipse polygon data with method label}
\usage{
create_ellipse_polygon_data(x_coords, y_coords, method, n_points = 100)
}
\description{
Create ellipse polygon data with method label
}
