% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{create_grid_cell_polygons}
\alias{create_grid_cell_polygons}
\title{Create grid cell polygons for plotting}
\usage{
create_grid_cell_polygons(center_x, center_y, grid_resolution)
}
\description{
Create grid cell polygons for plotting
}
