% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{create_grid_polygon}
\alias{create_grid_polygon}
\title{Create grid cell polygon}
\usage{
create_grid_polygon(x_coords, y_coords, grid_resolution)
}
\description{
Create grid cell polygon
}
