% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{create_grid_raster_data}
\alias{create_grid_raster_data}
\title{Create grid raster data using actual raster cell boundaries}
\usage{
create_grid_raster_data(track_points, reference_raster = NULL)
}
\description{
Create grid raster data using actual raster cell boundaries
}
