% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{create_space_use_polygon}
\alias{create_space_use_polygon}
\title{Create space use polygon for visualization}
\usage{
create_space_use_polygon(track_points, method)
}
\description{
Create space use polygon for visualization
}
