% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{create_space_use_polygon_data}
\alias{create_space_use_polygon_data}
\title{Create space use polygon data for direct plotting}
\usage{
create_space_use_polygon_data(x_coords, y_coords, method)
}
\description{
Create space use polygon data for direct plotting
}
