% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{create_space_use_summary}
\alias{create_space_use_summary}
\title{Create summary statistics for space use estimates}
\usage{
create_space_use_summary(space_use_df, methods, grouping_vars)
}
\description{
Create summary statistics for space use estimates
}
