% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_space_use_thresholds.R
\name{create_threshold_comparison_plots}
\alias{create_threshold_comparison_plots}
\title{Create visualization plots for threshold comparison}
\usage{
create_threshold_comparison_plots(
  area_comparison,
  track_overlap,
  plot_by_fish = TRUE
)
}
\description{
Create visualization plots for threshold comparison
}
