% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.12_function_space_use_absence_generation.R
\name{extract_absence_time_info}
\alias{extract_absence_time_info}
\title{Extract time binning information from space_use_points and space_use_results}
\usage{
extract_absence_time_info(space_use_points, space_use_results)
}
\description{
Extract time binning information from space_use_points and space_use_results
}
