% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.10_function_space_use_point_generation.R
\name{extract_space_use_time_info}
\alias{extract_space_use_time_info}
\title{Extract time binning information from space_use_results}
\usage{
extract_space_use_time_info(space_use_results)
}
\description{
Extract time binning information from space_use_results
}
