% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.12_function_space_use_absence_generation.R
\name{generate_absences_available_habitat}
\alias{generate_absences_available_habitat}
\title{Generate absences from available habitat (whole study area minus space use areas)}
\usage{
generate_absences_available_habitat(
  presence_points,
  space_use_results,
  track_data,
  fish_id,
  time_period,
  n_points,
  reference_raster,
  exclude_space_use,
  time_bin_info
)
}
\description{
Generate absences from available habitat (whole study area minus space use areas)
}
