% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.12_function_space_use_absence_generation.R
\name{generate_absences_buffer_zone}
\alias{generate_absences_buffer_zone}
\title{Generate absences in buffer zone around space use area}
\usage{
generate_absences_buffer_zone(
  presence_points,
  buffer_distance,
  n_points,
  reference_raster,
  exclude_space_use
)
}
\description{
Generate absences in buffer zone around space use area
}
