% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.12_function_space_use_absence_generation.R
\name{generate_absences_random_cells}
\alias{generate_absences_random_cells}
\title{Generate random absence points from any valid raster cells}
\usage{
generate_absences_random_cells(
  presence_points,
  n_points,
  reference_raster,
  exclude_space_use
)
}
\description{
Generate random absence points from any valid raster cells
}
