% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.12_function_space_use_absence_generation.R
\name{generate_absences_system_wide}
\alias{generate_absences_system_wide}
\title{Generate absences using system-wide detection efficiency}
\usage{
generate_absences_system_wide(
  presence_points,
  system_data,
  prob_threshold,
  n_points,
  reference_raster
)
}
\description{
Generate absences using system-wide detection efficiency
}
