% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.10_function_space_use_point_generation.R
\name{generate_points_bounding_box_method}
\alias{generate_points_bounding_box_method}
\title{Generate points within bounding box}
\usage{
generate_points_bounding_box_method(track_data, n_points, method = "random")
}
\description{
Generate points within bounding box
}
