% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.10_function_space_use_point_generation.R
\name{generate_points_constrained_hull}
\alias{generate_points_constrained_hull}
\title{Generate points within constrained convex hull}
\usage{
generate_points_constrained_hull(
  track_data,
  n_points,
  reference_raster = NULL,
  method = "random"
)
}
\description{
Generate points within constrained convex hull
}
