% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.10_function_space_use_point_generation.R
\name{generate_points_convex_hull}
\alias{generate_points_convex_hull}
\title{Generate points within convex hull polygon}
\usage{
generate_points_convex_hull(track_data, n_points, method = "random")
}
\description{
Generate points within convex hull polygon
}
