% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.10_function_space_use_point_generation.R
\name{generate_points_ellipse}
\alias{generate_points_ellipse}
\title{Generate points within 95\% confidence ellipse}
\usage{
generate_points_ellipse(track_data, n_points, method = "random")
}
\description{
Generate points within 95\% confidence ellipse
}
