% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.10_function_space_use_point_generation.R
\name{generate_points_grid_cells}
\alias{generate_points_grid_cells}
\title{Generate points within grid cells occupied by track}
\usage{
generate_points_grid_cells(
  track_data,
  n_points,
  reference_raster = NULL,
  method = "random"
)
}
\description{
Generate points within grid cells occupied by track
}
