% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.12_function_space_use_absence_generation.R
\name{generate_simple_raster_absences}
\alias{generate_simple_raster_absences}
\title{Simple absence generation function that ONLY uses the raster - no other inputs}
\usage{
generate_simple_raster_absences(reference_raster, n_points = 100)
}
\description{
Simple absence generation function that ONLY uses the raster - no other inputs
}
