% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.12_function_space_use_absence_generation.R
\name{generate_space_use_absences_simple}
\alias{generate_space_use_absences_simple}
\title{Wrapper to generate absences with fish/time info using simple raster logic}
\usage{
generate_space_use_absences_simple(
  space_use_points,
  reference_raster,
  n_points_ratio = 1
)
}
\description{
Wrapper to generate absences with fish/time info using simple raster logic
}
