% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.12_function_space_use_absence_generation.R
\name{get_space_use_exclusion_cells}
\alias{get_space_use_exclusion_cells}
\title{Get cells to exclude based on space use method}
\usage{
get_space_use_exclusion_cells(track_data, method, reference_raster)
}
\description{
Get cells to exclude based on space use method
}
