% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{line_intersects_box}
\alias{line_intersects_box}
\title{Check if a line segment intersects with a box
Uses basic line-box intersection algorithm}
\usage{
line_intersects_box(x1, y1, x2, y2, xmin, ymin, xmax, ymax)
}
\description{
Check if a line segment intersects with a box
Uses basic line-box intersection algorithm
}
