% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.0_function_particle_filter.R
\name{pf_position_estimate}
\alias{pf_position_estimate}
\title{Compute MAP or mean position estimate from weighted particles}
\usage{
pf_position_estimate(px, py, weights, method, raster)
}
\description{
Compute MAP or mean position estimate from weighted particles
}
\keyword{internal}
