% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.0_function_particle_filter.R
\name{pf_prepare_detections}
\alias{pf_prepare_detections}
\title{Standardize detection data column names}
\usage{
pf_prepare_detections(detection_data, fish_id_col, time_col, station_col)
}
\description{
Standardize detection data column names
}
\keyword{internal}
