% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.0_function_particle_filter.R
\name{pf_prepare_stations}
\alias{pf_prepare_stations}
\title{Prepare station data for particle filter}
\usage{
pf_prepare_stations(station_info)
}
\description{
Prepare station data for particle filter
}
\keyword{internal}
