% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.14_function_comparative_habitat_selection.R
\name{plot_comparative_results}
\alias{plot_comparative_results}
\title{Plot comparative habitat selection results}
\usage{
plot_comparative_results(comp_results, plot_type = "all")
}
\arguments{
\item{comp_results}{Output from analyze_comparative_habitat_selection}

\item{plot_type}{Type of plots to show ("individual", "comparison", "all")}
}
\description{
Plot comparative habitat selection results
}
