% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.9_function_habitat_selection_analysis.R
\name{plot_depth_spatial}
\alias{plot_depth_spatial}
\title{Create spatial plot of presence/absence points colored by depth}
\usage{
plot_depth_spatial(
  habitat_data,
  fish_select = 1,
  time_select = 0,
  background_raster = NULL,
  point_size = 1
)
}
\arguments{
\item{habitat_data}{Data frame with x, y, depth_m, type columns}

\item{fish_select}{Fish ID to plot (default: 1)}

\item{time_select}{Time period to plot (default: 0)}

\item{background_raster}{Optional background raster}

\item{point_size}{Size of points (default: 1)}
}
\value{
ggplot object
}
\description{
Create spatial plot of presence/absence points colored by depth
}
