% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.12_function_space_use_absence_generation.R
\name{plot_presence_absence_space_use}
\alias{plot_presence_absence_space_use}
\title{Plot presence and absence points together}
\usage{
plot_presence_absence_space_use(
  presence_points,
  absence_results,
  fish_select = NULL,
  time_select = NULL,
  background_raster = NULL,
  show_space_use = TRUE,
  track_data = NULL,
  space_use_results = NULL,
  reference_raster = NULL
)
}
\arguments{
\item{presence_points}{Output from generate_random_points_in_space_use}

\item{absence_results}{Output from generate_space_use_absences}

\item{fish_select}{Which fish to plot}

\item{time_select}{Which time period to plot}

\item{background_raster}{Optional background raster}

\item{show_space_use}{Whether to show space use area outline}

\item{track_data}{Optional track data to show}

\item{space_use_results}{Optional space use results for area visualization}

\item{reference_raster}{Optional reference raster for space use visualization}
}
\description{
Plot presence and absence points together
}
