% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{plot_space_use}
\alias{plot_space_use}
\title{Plot space use estimates}
\usage{
plot_space_use(
  space_use_results,
  method_compare = NULL,
  plot_type = "comparison",
  track_data = NULL,
  fish_select = NULL,
  time_select = NULL,
  method_select = "convex_hull",
  background_raster = NULL,
  time_aggregation = "day",
  point_type = "track"
)
}
\description{
Plot space use estimates
}
