% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{plot_space_use_map}
\alias{plot_space_use_map}
\title{Updated plot function with proper raster cell visualization}
\usage{
plot_space_use_map(
  space_use_results,
  track_data,
  fish_select,
  time_select,
  method_select,
  background_raster,
  time_aggregation,
  point_type = "track"
)
}
\description{
Updated plot function with proper raster cell visualization
}
