% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/1.7.11_function_space_use_point_generation_part2.R
\name{plot_space_use_method_comparison}
\alias{plot_space_use_method_comparison}
\title{Plot comparison of multiple space use methods}
\usage{
plot_space_use_method_comparison(
  comparison_points,
  track_data,
  fish_select = NULL,
  time_select = NULL,
  background_raster = NULL
)
}
\arguments{
\item{comparison_points}{Output from compare_space_use_methods}

\item{track_data}{Original track data}

\item{fish_select}{Which fish to plot}

\item{time_select}{Which time period to plot}

\item{background_raster}{Optional background raster}
}
\value{
A ggplot2 object comparing space use methods with generated points
}
\description{
Plot comparison of multiple space use methods
}
