% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.10_function_space_use_point_generation.R,
%   R/1.7.4_function_track_UD.R
\name{point_in_polygon}
\alias{point_in_polygon}
\title{Point-in-polygon test using ray casting algorithm}
\usage{
point_in_polygon(px, py, poly_x, poly_y)

point_in_polygon(px, py, poly_x, poly_y)
}
\description{
Point-in-polygon test using ray casting algorithm

Point-in-polygon test using ray casting algorithm
}
