% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.14_function_comparative_habitat_selection.R
\name{print_comparative_summary}
\alias{print_comparative_summary}
\title{Print comparative habitat selection summary}
\usage{
print_comparative_summary(comp_results)
}
\arguments{
\item{comp_results}{Output from analyze_comparative_habitat_selection}
}
\description{
Print comparative habitat selection summary
}
