% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.15_function_information_weighting.R
\name{print_station_characteristics}
\alias{print_station_characteristics}
\title{Print Station Characteristics Summary}
\usage{
print_station_characteristics(station_chars, n_stations = 10)
}
\arguments{
\item{station_chars}{Data frame from calculate_station_de_characteristics()}

\item{n_stations}{Number of stations to display (default 10)}
}
\description{
Prints a summary of station DE characteristics for diagnostic purposes
}
