% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.10_function_space_use_point_generation.R
\name{standardize_track_data_for_points}
\alias{standardize_track_data_for_points}
\title{Standardize track data for point generation to match space_use_results time binning}
\usage{
standardize_track_data_for_points(track_data, time_bin_info)
}
\description{
Standardize track data for point generation to match space_use_results time binning
}
