% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{standardize_track_data_space}
\alias{standardize_track_data_space}
\title{Standardize track data for space use analysis}
\usage{
standardize_track_data_space(track_data, time_aggregation)
}
\description{
Standardize track data for space use analysis
}
