% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/1.7.11_function_space_use_point_generation_part2.R
\name{summarize_space_use_points}
\alias{summarize_space_use_points}
\title{Summary function for space use points}
\usage{
summarize_space_use_points(space_use_points)
}
\arguments{
\item{space_use_points}{Output from generate_random_points_in_space_use}
}
\value{
A data frame with summary statistics for each fish and time period
}
\description{
Summary function for space use points
}
