% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.4_function_track_UD.R
\name{test_cells_in_polygon}
\alias{test_cells_in_polygon}
\title{Test if raster cells intersect with a polygon using simple corner test
Include cell if ANY of its 4 corners are inside the polygon}
\usage{
test_cells_in_polygon(raster_obj, cell_indices, poly_x, poly_y)
}
\description{
Test if raster cells intersect with a polygon using simple corner test
Include cell if ANY of its 4 corners are inside the polygon
}
