% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.7.12_function_space_use_absence_generation.R
\name{test_raster_absence_generation}
\alias{test_raster_absence_generation}
\title{Test function to validate raster and absence generation}
\usage{
test_raster_absence_generation(reference_raster, n_test_points = 10)
}
\description{
Test function to validate raster and absence generation
}
