% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.0_function_particle_filter.R
\name{pf_compute_single_ud}
\alias{pf_compute_single_ud}
\title{Compute single UD from weighted particles}
\usage{
pf_compute_single_ud(x, y, weights, raster, cell_area, contour_levels)
}
\description{
Compute single UD from weighted particles
}
\keyword{internal}
